Add mixin and update files
